+++ src/KOKKOS/npair_trim_kokkos.cpp
@@ -91,9 +91,9 @@ void NPairTrimKokkos<DeviceType>::operator()(TagNPairTrim, const int &ii) const 
   int n = 0;
 
   const int i = d_ilist_copy(ii);
-  const double xtmp = x(i,0);
-  const double ytmp = x(i,1);
-  const double ztmp = x(i,2);
+  const double xtmp = static_cast<double>(x(i,0));
+  const double ytmp = static_cast<double>(x(i,1));
+  const double ztmp = static_cast<double>(x(i,2));
 
   // loop over copy neighbor list
 
@@ -105,9 +105,9 @@ void NPairTrimKokkos<DeviceType>::operator()(TagNPairTrim, const int &ii) const 
     const int joriginal = d_neighbors_copy(i,jj);
     const int j = joriginal & NEIGHMASK;
 
-    const double delx = xtmp - x(j,0);
-    const double dely = ytmp - x(j,1);
-    const double delz = ztmp - x(j,2);
+    const double delx = xtmp - static_cast<double>(x(j,0));
+    const double dely = ytmp - static_cast<double>(x(j,1));
+    const double delz = ztmp - static_cast<double>(x(j,2));
     const double rsq = delx*delx + dely*dely + delz*delz;
 
     if (rsq > cutsq_custom) continue;
